[fix] Added more type and value check for Redis backend config file to prevent unconditional setting. (#159)
* More type and value check for config file and more reasonable exception catching mechanisms

Simplefy reading json and testing Redis service cluster_enabled and Redis connection parameters code by using macro and father class function

Feed all the exception catching mechanisms into tensorFlow's OP_REQUIRES mechanism and Add more CI about config file mistakes.

* Added a parameter redis_config_abs_dir_env in OP attribute which is the name of the environment variable that stores the path of the Redis configuration file to be read by the specified model for the operation of one machine and multiple models.

* Add redis_read_access_slave in config file to set if all reading access in Redis backend through SLAVE service in redis. In most infer mode sitiuation it could be set true.

Fix bug that's multi-thread throwing exception unable to be catched.

* fix delocate version to 0.9.1. delocate 0.10.0 makes delocate-wheel fail in MacOS

* Many functions were refactored(rebucket) and more bugs were fixed

modify compute_capabilities function in cuda_configure.bzl in case someone sets the environment variable to 11.x

+++ tensorflow_recommenders_addons/dynamic_embedding/python/ops/redis_table_ops.py
@@ -21,6 +21,8 @@ from __future__ import print_function
 import functools
 import json
 import os
+from re import T
+import warnings
 
 from tensorflow.python.eager import context
 from tensorflow.python.framework import dtypes
@@ -64,6 +66,8 @@ class RedisTable(LookupInterface):
       "redis_user": "default",
       "redis_password": "",
       "redis_db": 0,
+      "redis_read_access_slave":
+          False,  # set True in infer or train mode if you like
       "redis_connect_keep_alive": False,  # keep TCP alive
       "redis_connect_timeout": 1000,  # milliseconds
       "redis_socket_timeout": 1000,  # milliseconds
@@ -140,15 +144,68 @@ class RedisTable(LookupInterface):
     self._embedding_name = (self._name.split('_mht_', 1))[0]
     self._config = config
 
-    with open(self._config.redis_config_abs_dir, 'r', encoding='utf-8') as f0:
-      params_load = json.load(f0)
-      self._redis_params = self.default_redis_params.copy()
-      for k in self._redis_params.keys():
-        if k in params_load:
-          self._redis_params[k] = params_load[k]
+    self.redis_config_file_exist = False
+    self.redis_config_file_create = False
 
-    with open(self._config.redis_config_abs_dir, 'w', encoding='utf-8') as f1:
-      f1.write(json.dumps(self._redis_params, indent=2, ensure_ascii=True))
+    if self._config.redis_config_abs_dir_env:
+      if self._config.redis_config_abs_dir_env in os.environ:
+        self._config.redis_config_abs_dir = os.getenv(
+            self._config.redis_config_abs_dir_env)
+      else:
+        raise ValueError(
+            "Config redis_config_abs_dir_env in RedisTableConfig is not None, but can not find the "
+            + self._config.redis_config_abs_dir_env +
+            " in system environment variable.")
+      self.redis_config_file_exist = os.path.exists(
+          self._config.redis_config_abs_dir)
+      if self.redis_config_file_exist == False:
+        raise ValueError(
+            "Config redis_config_abs_dir_env in RedisTableConfig is not None, but the FILE which path stored in environment variable "
+            + self._config.redis_config_abs_dir_env + " DOES NOT EXIST.")
+    elif self._config.redis_config_abs_dir_env is None and "TFRA_REDIS_CONFIG_PATH" in os.environ:
+      self._config.redis_config_abs_dir = os.getenv("TFRA_REDIS_CONFIG_PATH")
+      self.redis_config_file_exist = os.path.exists(
+          self._config.redis_config_abs_dir)
+      warnings.warn(
+          "TFRA-Redis try to use environment variable TFRA_REDIS_CONFIG_PATH regardless redis_config_abs_dir in RedisTableConfig."
+      )
+      if self.redis_config_file_exist == False:
+        raise ValueError(
+            "environment variable TFRA_REDIS_CONFIG_PATH exists, but the FILE which path stored in TFRA_REDIS_CONFIG_PATH DOES NOT EXIST. Please create a FILE in the corresponding path or delete the environment variable TFRA_REDIS_CONFIG_PATH."
+        )
+    elif self._config.redis_config_abs_dir_env is None and "TFRA_REDIS_CONFIG_PATH" not in os.environ and self._config.redis_config_abs_dir:
+      self.redis_config_file_exist = os.path.exists(
+          self._config.redis_config_abs_dir)
+      if self.redis_config_file_exist == False:
+        raise ValueError(
+            "Config redis_config_abs_dir in RedisTableConfig is not None and redis_config_abs_dir_env is None, but the FILE "
+            + self._config.redis_config_abs_dir +
+            " which path is redis_config_abs_dir DOES NOT EXIST.")
+    elif self._config.redis_config_abs_dir_env is None and "TFRA_REDIS_CONFIG_PATH" not in os.environ and self._config.redis_config_abs_dir is None:
+      self.redis_config_file_create = True
+      self._config.redis_config_abs_dir = "/tmp/tmp_TFRA_Redis_config_file.json"
+      warnings.warn(
+          "Both redis_config_abs_dir_env and redis_config_abs_dir in RedisTableConfig are None, now creating a temporary config file in /tmp/tmp_TFRA_Redis_config_file.json."
+      )
+    else:
+      raise ValueError(
+          "TFRA-Redis didn't get the correct RedisTableConfig class initial parameter."
+      )
+
+    if self.redis_config_file_create == True and self.redis_config_file_exist == False:
+      with open(self._config.redis_config_abs_dir, 'w+',
+                encoding='utf-8') as f0:
+        f0.write(
+            json.dumps(self.default_redis_params, indent=2, ensure_ascii=True))
+    else:
+      with open(self._config.redis_config_abs_dir, 'r', encoding='utf-8') as f0:
+        params_load = json.load(f0)
+        self._redis_params = self.default_redis_params.copy()
+        for k in self._redis_params.keys():
+          if k in params_load:
+            self._redis_params[k] = params_load[k]
+      with open(self._config.redis_config_abs_dir, 'w', encoding='utf-8') as f1:
+        f1.write(json.dumps(self._redis_params, indent=2, ensure_ascii=True))
 
     self._shared_name = None
     if context.executing_eagerly():
@@ -185,7 +242,8 @@ class RedisTable(LookupInterface):
         value_dtype=self._value_dtype,
         value_shape=self._default_value.get_shape(),
         embedding_name=self._embedding_name,
-        redis_config_abs_dir=self._config.redis_config_abs_dir)
+        redis_config_abs_dir=self._config.redis_config_abs_dir,
+        redis_config_abs_dir_env=self._config.redis_config_abs_dir_env)
 
     if context.executing_eagerly():
       self._table_name = None
